Extract build finalization to job
`BuildReportJob` should stand in place of the GitHub interactions in
`Review` in #727 with the plan of using it in the build workers
controller. Also change Commenter to accept a `Payload` and create its
own `PullRequest` instance.

Up next will be implementing the `BuildWorkers` controller
that will check for build completion, persist violations, and enqueue
this job.

https://trello.com/c/CdExHIVN